% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/area.R
\docType{data}
\name{area}
\alias{area}
\title{This Is A One-line Title Describing area}
\format{area is a data frame with 50 rows and 3 variables:
\describe{
  \item{state}{state (name)}
  \item{land}{area(square miles)}
  \item{farm}{ares(square miles)}
}}
\source{
These data are from John Gillett
(\url{http://www.stat.wisc.edu/~jgillett/327-3/2/farmLandArea.csv})
}
\usage{
area
}
\description{
This is the Description paragraph. Area is a very useful data set.
}
\keyword{datasets}

