% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bonus.R
\name{base.function.most.methods}
\alias{base.function.most.methods}
\title{Find the base R generic function with the most methods}
\usage{
base.function.most.methods()
}
\value{
A list containing:
\item{base.function}{The name of the generic function with the most methods.}
\item{n}{The number of methods it has.}
}
\description{
This function scans all objects in the "base" package to find which generic function
has the largest number of S3 methods.
}
\examples{
\dontrun{
  base.function.most.methods()
}
}
