% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef.lad.R
\name{coef.lad}
\alias{coef.lad}
\title{coefficients of lad()}
\usage{
\method{coef}{lad}(object, ...)
}
\arguments{
\item{object}{(description of x; one line per parameter: Arguments section)}

\item{...}{further arguments passed to or from other methods}
}
\value{
coefficient
}
\description{
(Description paragraph)
}
\details{
(Details section)
}
\examples{
# (Examples section)


}

