\name{fdfmQQPlot}
\alias{fdfmQQPlot}

\title{QQ Plot}

\description{
Side-by-side QQ plots of the sample versus estimated quantiles.
}

\usage{
fdfmQQPlot(x, robustQQline = TRUE, ...)
}

\arguments{
  \item{x}{an \code{fdfm} object.}

  \item{robustQQline}{a logical value. If \code{TRUE} a line is robust fit for each plot.}

  \item{\dots}{additional arguments are passed to the plotting functions.}
}

\keyword{robust}


