\name{get.fit.models.database}
\alias{get.fit.models.database}

\title{
  Model Comparison Database
}

\description{
  Returns a list of comparable models.
}

\usage{
  get.fit.models.database()
}

\value{
  a list of comparable models for \code{fit.models}.
}

\seealso{
  \code{\link{fit.models}}
}

\examples{
  get.fit.models.database()
}

%\keyword{librobust}
\keyword{utilities}

