\name{key}
\alias{key}

\title{
  Plot Key
}

\description{
  Don't use this function.  It's only here because the lattice package doesn't provide a key function.  Use legend instead.
}

\usage{
  key(x, y, text, lines, corner = c(0, 1), cex = par("cex"), col = par("col"), lty = par("lty"), lwd = par("lwd"), pch = par("pch"), ...)
}

\arguments{
  \item{x}{x}

  \item{y}{y}

  \item{text}{text}

  \item{lines}{lines}

  \item{corner}{corner}

  \item{cex}{cex}

  \item{col}{col}

  \item{lty}{lty}

  \item{lwd}{lwd}

  \item{pch}{pch}

  \item{\dots}{dots}
}

\section{Side Effects}{
  Adds key to plot.
}

\keyword{aplot}
