\name{leuk.dat}
\alias{leuk.dat}
\docType{data}

\title{Leuk Data}

\description{
An exmaple data set for the misclassification fitter in glmRob.
}

\usage{data(leuk.dat)}

\format{
  A data frame with 33 observations on the following 3 variables.
  \describe{
    \item{\code{wbc}}{a numeric vector.}
    \item{\code{ag}}{a numeric vector.}
    \item{\code{y}}{a numeric vector.}
  }
}

\source{
Don't know - if you know please email the package maintainer.
}

\examples{
data(leuk.dat)
}

\keyword{datasets}


