\name{lmfmOverlaidQQPlot}
\alias{lmfmOverlaidQQPlot}

\title{Overlaid QQ Plot}

\description{
Overlaid residual qqplots for each of the models in an lmfm object.
}

\usage{
lmfmOverlaidQQPlot(x, ...)
}

\arguments{
  \item{x}{an lmfm object.}

  \item{\dots}{additional arguments are passed to the low-level plotting functions.}
}

\value{
\code{x} is invisibly returned.
}

\keyword{utilities}


