\name{lmfmOverlaidResDenPlot}
\alias{lmfmOverlaidResDenPlot}

\title{Overlaid Residual Density Plot}

\description{
Overlaid residual density plots for each of the models in an lmfm object.
}

\usage{
lmfmOverlaidResDenPlot(x, ...)
}

\arguments{
  \item{x}{an lmfm object.}

  \item{\dots}{additional arguments are passed to the low-level plotting functions.}
}

\value{
\code{x} is invisibly returned.
}

\keyword{hplot}


