\name{lmfmResKernDenPlot}
\alias{lmfmResKernDenPlot}

\title{Residual Kernel Density Plot}

\description{
Produces side-by-side residual kernel density plots from an lmfm object.
}

\usage{
lmfmResKernDenPlot(x, ...)
}

\arguments{
  \item{x}{an lmfm object.}

  \item{\dots}{additional arguments are passed to the low-level plotting functions.}
}

\value{
\code{x} is invisibly returned.
}

\keyword{hplot}


