\name{mallows.dat}
\alias{mallows.dat}
\docType{data}

\title{Mallows Data}

\description{
An exmaple data set for the mallows fitter in glmRob.
}

\usage{data(mallows.dat)}

\format{
  A data frame with 70 observations on the following 4 variables.
  \describe{
    \item{\code{y}}{a numeric vector.}
    \item{\code{a}}{a numeric vector.}
    \item{\code{b}}{a numeric vector.}
    \item{\code{c}}{a numeric vector.}
  }
}

\source{
Don't know - if you know please email the package maintainer.
}

\examples{
data(mallows.dat)
}

\keyword{datasets}


