% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.lad.R
\name{predict.lad}
\alias{predict.lad}
\title{Predict use object in lad}
\usage{
\method{predict}{lad}(object, new.x, ...)
}
\arguments{
\item{object}{a regression}

\item{new.x}{a vector}

\item{...}{further arguments passed to or from other methods}
}
\value{
fitted value
}
\description{
(Description paragraph)
}
\details{
(Details section)
}
\examples{
# (Examples section)


}

