% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.lad.R
\name{print.lad}
\alias{print.lad}
\title{Print coefficents in lad}
\usage{
\method{print}{lad}(x, ...)
}
\arguments{
\item{x}{(description of x; one line per parameter: Arguments section)}

\item{...}{further arguments passed to or from other methods}
}
\value{
the named coeffecients vector
}
\description{
(Description paragraph)
}
\details{
(Details section)
}
\examples{
# (Examples section)


}

