\name{splus.assign}
\alias{splus.assign}

\title{Assign Translator}

\description{
Converts the R assign format to the Splus assign format.
}

\usage{
splus.assign(asgn, term.labels)
}

\arguments{
  \item{asgn}{an assign attribute.}

  \item{term.labels}{a term.labels attribute.}
}

\value{
an S-Plus like assign list.
}

\keyword{utilities}


