% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{area}
\alias{area}
\title{Farm Land Area Data}
\format{
A data frame with appropriate columns.
}
\source{
\url{https://pages.stat.wisc.edu/~byang/305/2/farmLandArea.csv}
}
\usage{
area
}
\description{
Data about farm land area from farmLandArea.csv.
}
\keyword{datasets}
