% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/area.R
\docType{data}
\name{area}
\alias{area}
\title{Land and Farm Area of the U.S.  
This is a dataset for the land and farm area of the 50 states in the U.S.}
\format{area is a data frame with 50 rows and 3 variables:}
\source{
These data is from (\url{"http://www.stat.wisc.edu/~jgillett/327-3/2/farmLandArea.csv"})
}
\usage{
area
}
\description{
Land and Farm Area of the U.S.  
This is a dataset for the land and farm area of the 50 states in the U.S.
}
\keyword{datasets}
