% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/area.R
\docType{data}
\name{area}
\alias{area}
\title{Data of land area and farm area of all 50 US states}
\format{area is a data frame with 50 rows and 3 variables:
\describe{
\item{state}{US State}
\item{land}{state land area (acres)}
\item{farm}{state farm area (acres)}
}}
\source{
These data are from an experiment I did with my friend William
(\url{http://www.stat.wisc.edu/~jgillett/327-3/2/farmLandArea.csv})
at the Guiness Brewery in Dublin in 1908.
}
\usage{
area
}
\description{
Data containing the name of the state, its respective land area, and its respective farm area.
}
\keyword{datasets}
