% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/area.R
\docType{data}
\name{area}
\alias{area}
\title{This Is A One-line Title Describing area}
\format{x is a data frame with 3 rows and 2 variables:
\describe{
  \item{state}{state names in US}
  \item{land}{land area}
  \item{farm}{farm area in the state}
}}
\source{
These data are obtained from 
(\url{http://www.stat.wisc.edu/~jgillett/327-3/2/farmLandArea.csv"})
}
\usage{
area
}
\description{
This is the Description paragraph. x is my favorite data set. Blah, blah,
blah.
}
\keyword{datasets}

