% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/area.R
\docType{data}
\name{area}
\alias{area}
\title{Land and farm area for 50 states}
\format{
area is a data frame with 50 rows and 2 variables:
\describe{
\item{land}{land area (sq mi)}
\item{farm}{farm area (sq mi)}
}
}
\source{
These data are from WIKIPEDIA
(\url{https://simple.wikipedia.org/wiki/List_of_U.S._states_and_territories_by_area})
}
\usage{
area
}
\description{
\code{area} contains 50 states with their land and farm area.
}
\keyword{datasets}
