% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lad.R
\name{coef.lad}
\alias{coef.lad}
\title{Extract coefficients from lad objects}
\usage{
\method{coef}{lad}(object, ...)
}
\arguments{
\item{object}{A lad object.}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
Extract coefficients from lad objects
}
