% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef.lad.R
\name{coef.lad}
\alias{coef.lad}
\title{Coefficient vectors of the lad object}
\usage{
\method{coef}{lad}(object, ...)
}
\arguments{
\item{object}{A lad object}

\item{...}{further arguments passed to or from other methods}
}
\value{
A vector of intercept and the slope from the optimization result (Value section)
}
\description{
A method to return coefficient vectors of the lad object(Description paragraph)
}
\details{
The method is specifically for object lad and can especially be used in abline () (Details section)
}
\examples{
plot (area$land, area$farm)
abline (lm (area$farm~area$land), col = "black")
abline (coef(lad (x = area$land, y = area$farm)), col = "green")
legend (x = 0, y = 200000, c("fit linear model", "least absolute deviation"),lty = c(1,1), col = c(1, 3), cex = 0.5)
}
