% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef.R
\name{coef.lad}
\alias{coef.lad}
\title{Add least absolute deviations regression line}
\usage{
\method{coef}{lad}(x)
}
\arguments{
\item{x}{LAD class object}
}
\value{
\beta_0 and \beta_1 of LAD regression
}
\description{
This is a function allows ceofficients from the \code{lad} function to be plotted onto an existing plot.
}
\details{
Typical usage is \code{abline(results)}
}
\examples{
results <- lad(x = area$land, y = area$farm)
plot(x = area$land, y = area$farm)
abline(results)

coef(results)

}
