% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef.R
\name{coef.lad}
\alias{coef.lad}
\title{coef method for lad}
\usage{
\method{coef}{lad}(ladObj)
}
\arguments{
\item{ladObj}{lad Object}
}
\value{
coefficients    intercept and slope
}
\description{
return the named coeffcients vector
}
\details{
return the named coeffcients vector
}
\examples{
 area = read.csv("http://www.stat.wisc.edu/~jgillett/327-3/2/farmLandArea.csv")
 out <- lad(x=area$land, y=area$farm)
 coef(out)
 plot(x=area$land, y=area$farm)
 abline(coef(out))
}

