% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef.lad.R
\name{coef.lad}
\alias{coef.lad}
\title{Return the coefficients}
\usage{
\method{coef}{lad}(object, ...)
}
\arguments{
\item{object}{a list of class \code{lad} given by \code{lad()}}

\item{...}{further arguments passed to or from other methods}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{a vector of coefficients without names
}\if{html}{\out{</div>}}
}
\description{
Return a vector of coefficients of a
least absolute deviations regression.
}
\details{
\code{coef.lad} gives a vector of coefficients without names
from a list of class \code{lad}.
}
