\name{gammaRob.control}
\alias{gammaRob.control}

\title{Control Parameters for gammaRob}

\description{
Create a list of control parameters for the \code{gammaRob} function.
}

\usage{
gammaRob.control(estim, \dots)
}

\arguments{
  \item{estim}{a character string specifying the estimator.}

  \item{\dots}{control parameters appropriate for the estimator given in \code{estim}.}
}

\value{
  a list of control parameters appropriate for the specified estimator.
}


\keyword{robust}


