% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lad.R
\name{lad}
\alias{lad}
\title{Least Absolute Deviation Regression}
\usage{
lad(beta, x, y)
}
\arguments{
\item{beta}{A numeric vector}

\item{x}{A numeric vector}

\item{y}{A numeric vector}
}
\value{
A list of coefficients, fitted values, and residuals from the optimization of least absolute square
}
\description{
Use the Nelder-Mead method to find the minimum absolute square of the data
}
\details{
The returned list is a list consists of three vectors, including coefficients which returns the optimized intercept and slop, fitted value which returns the predict value that can create least absolute deviation, and a residual vector
}
\examples{
lad(x= area$land, y = area$farm)

}
