% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lad.R
\name{lad}
\alias{lad}
\title{Least absolute deviations (LAD) regression}
\usage{
lad(x, y)
}
\arguments{
\item{x}{The predictor variable of a data frame used for regression.}

\item{y}{The response variable of a data fro mused for regression.}
}
\value{
Returns a list containing regression coefficients, fitted values, and residuals.
}
\description{
This function performs and optimization using the Nelder-Mead method to find the regression coefficients of two objects. 
This method minimizes the sum of absolution deviations of data X from predictions of Y.
}
\examples{
results <- lad(x = area$land, y = area$farm)
print(results)

plot(x = area$land, y = area$farm)
abline(lm(area$farm ~ area$land), col = "green")
abline(results, col = "blue")
legend("bottomright", legend = c("lm", "LAD"), lty = c(1, 1), col = c("green", "blue"))

}
