% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lad.R
\name{lad}
\alias{lad}
\title{Calculate the Least absolute deviation (title)}
\usage{
lad(x, y)
}
\arguments{
\item{x}{single predictor variable}

\item{y}{response variable}
}
\value{
coefficients   intercept and slope of the fitted line
        fitted.values  fitted.values
        residuals      rediduals
}
\description{
Calculate the Least absolute deviation for single covariate ("Description" paragraph)
}
\details{
(Details section)
}
\examples{
 # Example
 area = read.csv("http://www.stat.wisc.edu/~jgillett/327-3/2/farmLandArea.csv")
 out <- lad(x=area$land, y=area$farm)
 out
}

