% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lad.R
\name{lad}
\alias{lad}
\title{Least absolute deviations regression}
\usage{
lad(x, y)
}
\arguments{
\item{x}{a numeric vector}

\item{y}{a numeric vector with the same length as \code{x}}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{A list (\code{coefficients}, \code{fitted.values}, \code{residuals})
        whose class is "lad".
}\if{html}{\out{</div>}}
}
\description{
Do least absolute deviations regression.
}
\details{
\code{lad} uses \code{optim()} with its (default) Nelder-Mead
algorithm to minimize the sum of absolute deviations (SAD)
Initial values come from \code{lm()}
}
\examples{
data(area)
lad1 = lad(x=area$land, y=area$farm)
print(unname(lad1$coefficients))
plot(x=area$land, y=area$farm)
abline(lm(area$farm ~ area$land))
abline(lad1, col = "red")
legend(0, 200000, legend = c("lm", "lad"),
                  lty = c("solid", "solid"),
                  col = c("black", "red"))
quantiles <- quantile(area$land, probs = c(0, 0.25, 0.5, 0.75, 1))
quantiles <- unname(quantiles)
y_hat <- predict(lad1, quantiles)
points(quantiles, y_hat, col = "green", pch = 16)
}
