\name{lmfmOverlaidQQPlot}
\alias{lmfmOverlaidQQPlot}

\title{Overlaid Normal QQ Plot of Residuals}

\description{
Produces an overlaid plot comparing residuals QQ plots.
}

\usage{
lmfmOverlaidQQPlot(x, main, xlab, ylab, ...)
}

\arguments{
  \item{x}{an lmfm object.}

  \item{main}{a character string specifying the plot title.}

  \item{xlab}{a character string specifying the x-axis label.}

  \item{ylab}{a character string specifying the y-axis label.}

  \item{\dots}{additional arguments are passed to the low-level plotting functions.}
}

\value{
\code{x} is invisibly returned.
}

\keyword{hplot}


