\name{lmfmOverlaidResDenPlot}
\alias{lmfmOverlaidResDenPlot}

\title{Overlaid Residual Density Plot}

\description{
Overlaid residual density plots for each of the models in an lmfm object.
}

\usage{
lmfmOverlaidResDenPlot(x, main, xlab, ylab, ...)
}

\arguments{
  \item{x}{an lmfm object.}

  \item{main}{a character string specifying the plot title.}

  \item{xlab}{a character string specifying the x-axis label.}

  \item{ylab}{a character string specifying the y-axis label.}

  \item{\dots}{additional arguments are passed to the low-level plotting functions.}
}

\value{
\code{x} is invisibly returned.
}

\keyword{hplot}


