% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lad.R
\name{predict.lad}
\alias{predict.lad}
\title{Predict method for lad objects}
\usage{
\method{predict}{lad}(object, new.x, ...)
}
\arguments{
\item{object}{A lad object.}

\item{new.x}{A numeric vector of new x values to predict at.}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
Predict method for lad objects
}
