% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.lad.R
\name{predict.lad}
\alias{predict.lad}
\title{Predict optimized y value}
\usage{
\method{predict}{lad}(object, new.x, ...)
}
\arguments{
\item{object}{A lad object}

\item{new.x}{A number vector}

\item{...}{further arguments passed to or from other methods}
}
\value{
A vector of predicted value
}
\description{
A function that generate predicted value according to lad optimization
}
\details{
Enter a vector of independent values and a lad object, the predict.lad function will return predicted values according to the optimized least absolute deviation regression
}
\examples{
plot (area$land, area$farm)
points( x = quantile(area$land), y = predict(lad (x = area$land, y = area$farm), quantile(area$land)), col = "green" )

}
