% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.lad}
\alias{predict.lad}
\title{Predict function for LAD class object}
\usage{
\method{predict}{lad}(x, new.x = FALSE)
}
\arguments{
\item{x}{LAD class object}

\item{new.x}{Vector of new x values to generate a vector of new Y_hat values}
}
\value{
Y_hat A vector of predicted values using coefficients from LAD.
}
\description{
Written to explicitely on objects from the least absolute deviations (LAD) regression fucntion.
}
\examples{
results <- lad(x = area$land, y = area$farm)
predict(results)

predict(results, new.x = 1:10)

}
