% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.lad}
\alias{predict.lad}
\title{predict method for lad}
\usage{
\method{predict}{lad}(ladObj, newX)
}
\arguments{
\item{ladObj}{lad Object}

\item{newX}{new values of the covariate}
}
\value{
Ypred    predicted value at new values
}
\description{
predict the response at newly observed value of covariates
}
\details{
predict the response at newly observed value of covariates
}
\examples{
 area = read.csv("http://www.stat.wisc.edu/~jgillett/327-3/2/farmLandArea.csv")
 out <- lad(x=area$land, y=area$farm)
 ypred <- predict(out, newX=c(10000,20000))
}

