% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.lad.R
\name{predict.lad}
\alias{predict.lad}
\title{Lad's predictions}
\usage{
\method{predict}{lad}(object, new.x, ...)
}
\arguments{
\item{object}{a list of class \code{lad} given by \code{lad()}}

\item{new.x}{a vector of new values for \code{x}}

\item{...}{further arguments passed to or from other methods}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{lad's predictions based on \code{new.x}
}\if{html}{\out{</div>}}
}
\description{
Return a vector containing lad's predictions at the \code{x} values in \code{new.x}.
least absolute deviations regression.
}
\details{
\code{predict.lad} gives a vector containing lad's predictions
at the \code{x} values in \code{new.x}.
}
