% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lad.R
\name{print.lad}
\alias{print.lad}
\title{Print method for lad objects}
\usage{
\method{print}{lad}(x, ...)
}
\arguments{
\item{x}{A lad object.}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
Print method for lad objects
}
