% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.lad.R
\name{print.lad}
\alias{print.lad}
\title{A method to print lad object}
\usage{
\method{print}{lad}(x, ...)
}
\arguments{
\item{x}{A lad object}

\item{...}{further arguments passed to or from other methods}
}
\value{
A list of coefficients, fitted value, and residuals.
}
\description{
This is a function for creating print method for lad object
}
\details{
Return Intercept = "", x = ""
}
\examples{
a = lad (x = area$land, y = area$farm)
print(a)

}
