% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.lad}
\alias{print.lad}
\title{Print LAD object coefficients}
\usage{
\method{print}{lad}(x, ...)
}
\arguments{
\item{x}{An object from the \code{lad} function.}

\item{...}{Additional arguments.}
}
\value{
Regression coefficients to the console.
}
\description{
This is a function to print the regression coefficients from the \code{lad} function.
}
\examples{
results <- lad(x = area$land, y = area$farm)
print(results)

}
