% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.lad}
\alias{print.lad}
\title{print method}
\usage{
\method{print}{lad}(ladObj)
}
\arguments{
\item{ladObj}{lad Object}
}
\value{
NULL    nothing
}
\description{
writes the named coeffcients vector
}
\details{
writes the named coeffcients vector
}
\examples{

 area = read.csv("http://www.stat.wisc.edu/~jgillett/327-3/2/farmLandArea.csv")
 out <- lad(x=area$land, y=area$farm)
 print(out)
}

