% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.lad.R
\name{print.lad}
\alias{print.lad}
\title{Return the named coefficients}
\usage{
\method{print}{lad}(x, ...)
}
\arguments{
\item{x}{a list of class \code{lad} given by \code{lad()}}

\item{...}{further arguments passed to or from other methods}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{the named coefficients vector
}\if{html}{\out{</div>}}
}
\description{
Return the named coefficients of a
least absolute deviations regression.
}
\details{
\code{print.lad} gives the named coefficients vector
from a list of class \code{lad}.
}
